+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Option to exclude groups which are located outside group unit ([#33](https://github.com/scm-manager/scm-ldap-plugin/pull/33))
 ### Changed
 - Use custom ssl context ([#31](https://github.com/scm-manager/scm-ldap-plugin/pull/31))